finished up descriptions, updated item data
Finished the plural_description and single_description changes. Updated the data files to use this, so they're a lot less cluttered. Materials still use the old way since a material's description varies from item to item.

Also made most of the data files a bit more readable - Ammunition and scroll effects are split to their own files, and effects have had their IDs become less of a mash of letters.

Gave curses names to use in the UI, too. Not used anywhere else, but it looks more consistent this way.

Also removed setting item_description as a single thing, since it might change (particularly with ammo). So now it'll grab the description when it's needed.