First pass at system folders
There's an option to run locally, for those that want it.
Organized the data files into a data folder.
re-wrote configs.py to be able to use system folders.
For Mac OS X, it should be ~/Library/Application Support/nhg
For Unix, it should be ~/.local/share/nhg
For Windows it should be %appdata%\Local\andon\nhg