Configs update
The configs.py file now creates the classes that are then used by each location. If I remember how this works, this means one thing updating a file will update it for all of them. Not that that's something the program worries about currently.

Also adds memory to the last saved folder.